refactor(domain): add `services`, `valueObjects` and `entities` dir
