+++ src/application/usecases/GetAccount.ts
@@ -1,4 +1,3 @@
-import { Account } from "@/domain/Account";
 import { AccountRepository } from "@/infra/repositories/AccountRepository";
 
 export class GetAccountById {
