+++ src/domain/entities/Position.ts
@@ -1,0 +1,1 @@
+export class Position {}
