+++ src/domain/valueObjects/CarPlate.ts
