+++ src/domain/valueObjects/Coord.ts
