+++ src/infra/repositories/AccountRepositoryDatabase.ts
@@ -1,6 +1,6 @@
 import { AccountRepository } from "@/infra/repositories/AccountRepository";
-import { Account } from "@/domain/Account";
 import { DatabaseConnection } from "@/infra/database/DatabaseConnection";
+import { Account } from "@/domain/entities/Account";
 
 export class AccountRepositoryDatabase implements AccountRepository {
   constructor(private readonly connection: DatabaseConnection) {}
