+++ src/infra/repositories/PositionRepository.ts
@@ -1,0 +1,3 @@
+export interface PositionRepository {
+  save(position: Position): Promise<void>;
+}
